Базовая настройка стилей
