+++ config/serve.json
@@ -2,9 +2,9 @@
   "$schema": "https://developer.microsoft.com/json-schemas/core-build/serve.schema.json",
   "port": 4321,
   "https": true,
-  "initialPage": "https://localhost:5432/workbench",
+  "initialPage": "https://localhost:5433/workbench",
   "api": {
-    "port": 5432,
+    "port": 5433,
     "entryPath": "node_modules/@microsoft/sp-webpart-workbench/lib/api/"
   }
 }
