+++ src/webparts/rcrCalendar/components/RcrCalendar.module.scss
@@ -73,7 +73,7 @@
     }
   }
   .content {
-    width: auto;
+    width: 100%;
   }
 
   .dashboard {
